Предрелизная подготовка - перевод новых ресурсов на основной эндпойнт, смена нейминга старых ресурсов.
