+++ src/main/java/com/nekromant/telegram/controller/PricingController.java
@@ -5,8 +5,19 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class PricingController {
+
+    @GetMapping
+    public String redirectToPricing() {
+        return "redirect:/pricing";
+    }
+
     @GetMapping("/pricing")
-    public String getResumeAnalysisRequestPage() {
+    public String getPricingPage() {
         return "pricing";
+    }
+
+    @GetMapping("/pricingv1")
+    public String getResumeAnalysisRequestPage() {
+        return "pricing_old";
     }
 }
