+++ src/main/resources/static/css/pricing_old.css
@@ -1,0 +1,28 @@
+.no-space-break {
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    overflow: hidden;
+}
+
+.offer-checkbox {
+    padding: 0%;
+}
+
+.offer-list {
+    list-style-type: none;
+    padding: 0;
+}
+
+.marker {
+    display: inline-block;
+    width: 15px;
+    height: 15px;
+    margin-right: 10px;
+}
+
+.offer-resume-header {
+    padding: 0;
+    font-style: italic;
+    font-size: large;
+    color: darkgreen;
+}
